ethereum: Don't use cached block on `is_on_main_chain`
The block stream relies on this method checking with the Ethereum
node if a block is on the main chain, it is incorrect to assume
that a block in the cache is final if it is past the reorg threshold.
We are currently assuming that, as long as it is present
the block cache, the current subgraph head is final when
passed into `is_on_main_chain`, as long as it is present
in the block cache, which is incorrect.

The most likely way for this bug to affect a subgraph is if a large
reorg happens, such that by the time the last blocks are having the
reorged processed by graph node they have fallen out of the threshold
and are passed into `is_on_main_chain` and considered final. Since
mainnet does not see large reorgs relative to the threshold,
this should not affect subgraphs currently on the network.
In the hosted service, I have seen a Ropsten subgraph affected during
a 200 block reorg, which was how I became aware of this bug.