store: Map fdw tables without accessing the foreign server
When we initialize a connection pool, we should only assume that the
database for that pool is up and running. Previously, we used `import
foreign schema` to create fdw tables locally, but that requires that
Postgres access the remote server, and will therefore not work when that
server is not up.

We now create these tables by assuming the corresponding table on the
remote server has the same schema as our local table, which avoids
accessing the remote server. It also removes a subtle race where we could
have pulled in the schema from a remote server before that ran its latest
migrations.