store: Fix behavior of redeploy when deployment rules changed
When the deployment rules changed, and an already existing deployment was
deployed again, we would still create the subgraph schema in a different
shard. This additional data won't be used and not do any harm, except if
that deployment was used as the base for a graft where it would confuse a
safety check of the copying logic.

With these changes, we ignore deployment rules when asked to create an
already existing deployment and follow the decisions that were previously
made.