store: Optimize queries of child_type_d
Queries of type `child_type_d` get a list of `(parent_id, child_id)`
pairs. The parent id is only needed for bookkeeping when we process the
result, but does not constraint the result set. If there are only a few
unique `child_id`, Postgres would still plan this query as if all the
`child_id` were unique, which leads to bad query plans.

We now add a constraint `c.id = any({ unique child ids })` to help the
Postgres planner determine that there are only a few child ids and improve
its plan using that information.