graph: Improve weight estimation of BTreeMap
This commit also makes the stress example easier to handle, and copies the
new weight code from `cache_weight.rs` to make comparisons between the two
easier. Right now, both implementations are the same.

With the new estimation, BTreeMap's CacheWeight is precise for small maps
and closer to the true allocation for larger maps with more than ~ 22
entries.