Avoid checking chain continuity with subgraph block ptr if subgraph never processed any block (#3354)
FirehoseBlockStream needs to perform a chain continuity check when starting from a subgraph's existing block ptr because we cannot be sure to respect chain continuity when switching from RPC syncing to Firehose syncing in Ethereum.

However, the previous logic to check this was a little bit flawed. Indeed, the assumption that a `None` subgraph current block would mean the subgraph was never synced never actually match. Indeed, at deployment, even before the subgraph has synced something, the subgraph current block is set to the parent of the manifest's start block number. This means we were always doing the check if the subgraph was freshly deployed.

We now changed the logic. If the subgraph ptr is set but is < than the manifest start block num, it means the subgraph never synced anything in which case, we shall skip the continuity check. That will temporarily solve the Firehose Filters problem when deploying a fresh new Firehose subgraph with filters. It will also speed up starting subgraph on other chains since we will completely skip the continuity check there since it's not required.

Co-authored-by: Matthieu Vachon <matt@streamingfast.io>