chain/ethereum: Don't panic when failing to decode Firehose blocks
Before, graph-node would crash completely whenever it would encounter a
Firehose block that didn't meet the expectations. In the specific
example where we ran into this, there were invalid Firehose blocks
delivered to graph-node with `null` addresses that were expected to
be non-null.

This change makes it so that decoding errors are handled more
gracefully: they simply make the subgraph in question enter a retry loop
that recovers as soon as the Firehose issue is fixed. We could also fail
the subgraph but that's less robust.

Fixes #3460.